# apps/gateway-api/swagger.yaml
openapi: 3.0.3
info:
  title: FoodBot Gateway API
  description: |
    RESTful JSON API for the FoodBot restaurant commerce platform.
    Provides authentication, restaurant management, order processing,
    payment handling, and admin operations.
  version: 1.0.0
  contact:
    name: FoodBot Platform Team
    email: support@foodbot.com
servers:
  - url: http://localhost:3000
    description: Development server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/login or /auth/register
  schemas:
...
    Restaurant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        cuisineTypes:
          type: array
          items:
            type: string
        rating:
          type: number
        priceRange:
          type: string
        address:
          type: string
        location:
          type: object
          properties:
            latitude:
              type: number
            longitude:
              type: number
        operatingHours:
          type: object
          properties:
            open:
              type: string
            close:
              type: string
        isOpen:
          type: boolean
        ownerId:
          type: string
    Dish:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        price:
          type: number
        category:
          type: string
        restaurantId:
          type: string
        isVegetarian:
          type: boolean
        isVegan:
          type: boolean
        isAvailable:
          type: boolean
        preparationTime:
          type: integer
        nutritionalInfo:
          type: object
          properties:
            calories:
              type: integer
            protein:
              type: integer
            carbs:
              type: integer
            fat:
              type: integer
...
paths:
...
  /restaurants/search:
    get:
      tags: [Restaurants]
      summary: Search restaurants
      security: []
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: latitude
          in: query
          schema:
            type: number
        - name: longitude
          in: query
          schema:
            type: number
        - name: radius
          in: query
          schema:
            type: number
        - name: cuisineTypes
          in: query
          schema:
            type: array
            items:
              type: string
        - name: minRating
          in: query
          schema:
            type: number
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/Restaurant'
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
  /restaurants/{id}:
    get:
      tags: [Restaurants]
      summary: Get restaurant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Restaurant details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Restaurant'
        '404':
          description: Restaurant not found
...
  /restaurants:
    post:
      tags: [Restaurants]
      summary: Create a restaurant
      description: Requires restaurant_owner or admin role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, address, cuisineTypes]
              properties:
                name:
                  type: string
                description:
                  type: string
                cuisineTypes:
                  type: array
                  items:
                    type: string
                address:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
                priceRange:
                  type: string
                operatingHours:
                  type: object
      responses:
        '201':
          description: Restaurant created
        '403':
          description: Insufficient permissions
  /dishes/search:
    get:
      tags: [Dishes]
      summary: Search dishes
      security: []
      parameters:
        - name: query
          in: query
          schema:
            type: string
        - name: restaurantId
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - name: isVegetarian
          in: query
          schema:
            type: boolean
        - name: minPrice
          in: query
          schema:
            type: number
        - name: maxPrice
          in: query
          schema:
            type: number
      responses:
        '200':
          description: Dish search results
  /dishes/{id}:
    get:
      tags: [Dishes]
      summary: Get dish by ID
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Dish details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Dish'
  /dishes:
    post:
      tags: [Dishes]
      summary: Create a dish
      description: Requires restaurant_owner or admin role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Dish'
      responses:
        '201':
          description: Dish created
...